requirements.txt
absl-py==0.13.0
asgiref==3.2.10
astor==0.8.1
cached-property==1.5.2
Django==3.1.2
gast==0.4.0
google-pasta==0.2.0
gunicorn==20.1.0
importlib-metadata==4.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
Pillow==8.0.1
protobuf==3.17.3
pytz==2021.1
sqlparse==0.4.1
tensorflow-cpu==2.5.0
termcolor==1.1.0
Werkzeug==2.0.1
zipp==3.4.1
h5py==3.1.0
numpy==1.19.2
six==1.15.0
whitenoise==5.3.0
django-heroku